Added ApiError, Runtime exception for testing. Moved NotFoundException to exception package. Added response.status for creation of new entry
