+++ src/main/java/com/teamtreehouse/blog/exception/ApiError.java
@@ -1,0 +1,13 @@
+package com.teamtreehouse.blog.exception;
+
+public class ApiError extends RuntimeException {
+    private final int status;
+    public ApiError(int status, String msg) {
+        super(msg);
+        this.status = status;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+}
